requirements.txt
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
streamlit==1.29.0
pandas==2.1.4
plotly==5.18.0
python-dateutil==2.8.2
